Pull more helpers out of gatsby-node
