+++ src/utils/setup/construct-db-filter.js
@@ -1,0 +1,6 @@
+const constructDbFilter = pageIdPrefix => ({
+    page_id: { $regex: new RegExp(`^${pageIdPrefix}/*`) },
+    commit_hash: process.env.COMMIT_HASH || { $exists: false },
+});
+
+module.exports = { constructDbFilter };
