+++ src/utils/setup/save-asset-files.js
@@ -1,0 +1,42 @@
+const fs = require('fs').promises;
+const mkdirp = require('mkdirp');
+const path = require('path');
+const { ASSETS_COLLECTION } = require('../../build-constants');
+const { getMetadata } = require('../get-metadata');
+
+const metadata = getMetadata();
+const DB = metadata.database;
+
+const saveAssetFile = async asset => {
+    return new Promise((resolve, reject) => {
+        // Create nested directories as specified by the asset filenames if they do not exist
+        mkdirp(path.join('static', path.dirname(asset.filename)), err => {
+            if (err) return reject(err);
+            fs.writeFile(
+                path.join('static', asset.filename),
+                asset.data.buffer,
+                'binary',
+                err => {
+                    if (err) reject(err);
+                }
+            );
+            resolve();
+        });
+    });
+};
+
+// Write all assets to static directory
+const saveAssetFiles = async (assets, stitchClient) => {
+    const promises = [];
+    const assetQuery = { _id: { $in: assets } };
+    const assetDataDocuments = await stitchClient.callFunction(
+        'fetchDocuments',
+        [DB, ASSETS_COLLECTION, assetQuery]
+    );
+    assetDataDocuments.forEach(asset => {
+        promises.push(saveAssetFile(asset));
+    });
+    return Promise.all(promises);
+};
+
+module.exports = { saveAssetFiles };
